+++ plugins/ecdsa/index.ts
@@ -21,3 +21,12 @@ export {
     type CreateEcdsaKernelMigrationAccountParameters,
     type CreateEcdsaKernelMigrationAccountReturnType
 } from "./account/createEcdsaKernelMigrationAccount.js"
+export {
+    create7702KernelAccountClient,
+    type Create7702KernelAccountClientParameters
+} from "./clients/kernel7702AccountClient.js"
+export {
+    create7702KernelAccount,
+    type Create7702KernelAccountParameters,
+    type Create7702KernelAccountReturnType
+} from "./account/create7702KernelAccount.js"