+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @zerodev/sdk
 
+## 5.4.30
+
+### Patch Changes
+
+- client action extends issue fixed
+
+## 5.4.29
+
+### Patch Changes
+
+- updated to use latest viem 2.28.0 and support for 7702
+
 ## 5.4.28
 
 ### Patch Changes