+++ plugins/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/hooks
 
+## 5.3.4
+
+### Patch Changes
+
+- updated to use latest viem 2.28.0 and support for 7702
+
 ## 5.3.3
 
 ### Patch Changes