+++ plugins/passkey/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/passkey-validator
 
+## 5.5.4
+
+### Patch Changes
+
+- updated to use latest viem 2.28.0 and support for 7702
+
 ## 5.5.3
 
 ### Patch Changes