+++ plugins/permission/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/permissions
 
+## 5.5.5
+
+### Patch Changes
+
+- updated to use latest viem 2.28.0 and support for 7702
+
 ## 5.5.4
 
 ### Patch Changes