+++ plugins/weighted-r1-k1/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zerodev/weighted-validator
 
+## 5.4.6
+
+### Patch Changes
+
+- updated to use latest viem 2.28.0 and support for 7702
+
 ## 5.4.5
 
 ### Patch Changes